[sway] Add screenshot shortcut
$mod+Shift+S: let use choose region, and take a screenshot at that
region and put it into the clipboard.

The region is predefined as the different windows that are visible but
the user can override that with their own selection.

I was told on #sway:irc.freenode.net that it was easiest to put the
screenshot command inside of a script rather than trying to escape it.